requirements.txt
# ── Frontend ──────────────────────────────────
streamlit==1.35.0

# ── Backend ───────────────────────────────────
fastapi==0.110.0
uvicorn==0.30.1
typer==0.9.4
python-multipart==0.0.9

# ── Auth ──────────────────────────────────────
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.3

# ── Database ──────────────────────────────────
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9

# ── Validation ────────────────────────────────
pydantic==2.7.4
pydantic-settings==2.3.4
email-validator==2.1.1

# ── Scraping ──────────────────────────────────
requests==2.31.0
beautifulsoup4==4.12.3
playwright==1.44.0
lxml==5.2.1

# ── LLM ───────────────────────────────────────
groq==0.9.0
langchain==0.2.5
langchain-groq==0.1.6
langgraph==0.1.9

# ── Resume ────────────────────────────────────
pdfplumber==0.11.0
PyMuPDF==1.24.5
pyresparser==1.0.6
reportlab==4.2.0
python-docx==1.1.2

# ── NLP + Scoring ─────────────────────────────
spacy==3.7.4
sentence-transformers==3.0.1
scikit-learn==1.5.0
numpy==1.26.4

# ── Email ─────────────────────────────────────
dnspython==2.6.1
pyIsEmail==1.1.0
imapclient==3.0.1

# ── Google Sheets ─────────────────────────────
gspread==6.1.2
google-auth==2.30.0
google-auth-oauthlib==1.2.0

# ── Scheduling ────────────────────────────────
APScheduler==3.10.4

# ── Utils ─────────────────────────────────────
python-dotenv==1.0.1
loguru==0.7.2
tqdm==4.66.4
python-dateutil==2.9.0
duckduckgo-search==6.2.6